check if doc generation is successed; add commit id to the message when auto deploying

+++ code_docs.sh
@@ -9,16 +9,21 @@ rm -rf code_docs/*
 # generate documentation
 doxygen docs/Doxyfile
 
-# move documentation files to gh-pages branch
-mv docs/html/* code_docs/
-
-cd code_docs
-
-# configure git
-git config user.name "stevenliuyi"
-git config user.email "stevenliuyi@gmail.com"
-
-# add documentations and push back to gh-pages
-git add --all
-git commit -m "Doxygen deployment - Travis build: ${TRAVIS_BUILD_NUMBER}"
-git push https://${GITHUB_TOKEN}@github.com/stevenliuyi/lcs gh-pages:gh-pages --force
+# check if documentation generation is successed
+if [ -d "docs/html" ] && [ -f "docs/html/index.html" ]; then
+    # move documentation files to gh-pages branch
+    mv docs/html/* code_docs/
+    
+    cd code_docs
+    
+    # configure git
+    git config user.name "stevenliuyi"
+    git config user.email "stevenliuyi@gmail.com"
+    
+    # add documentations and push back to gh-pages
+    git add --all
+    git commit -m "Doxygen deployment - Travis build: ${TRAVIS_BUILD_NUMBER} Commit: ${TRAVIS_COMMIT}"
+    git push https://${GITHUB_TOKEN}@github.com/stevenliuyi/lcs gh-pages:gh-pages --force
+else
+    exit 1
+fi
